move CaptionSplit to new js file
